Add flag to run only host-network backed pods

+++ cleanup.sh
@@ -1,13 +1,66 @@
 #!/bin/bash
 
+# source the functions in labels.sh
+. labels.sh
+
 #
 # Default values (possible to override)
 #
 CLEAN_ALL=${CLEAN_ALL:-false}
+FT_HOSTONLY=${FT_HOSTONLY:-unknown}
+HTTP_SERVER_POD_NAME=${HTTP_SERVER_POD_NAME:-ft-http-server-pod-v4}
 
+dump-working-data() {
+  echo
+  echo "Default/Override Values:"
+  echo "  Launch Control:"
+  echo "    FT_HOSTONLY                        $FT_HOSTONLY"
+  echo "    HTTP_SERVER_POD_NAME               $HTTP_SERVER_POD_NAME"
+  echo "    CLEAN_ALL                          $CLEAN_ALL"
+  echo "    FT_REQ_SERVER_NODE                 $FT_REQ_SERVER_NODE"
+  echo "    FT_REQ_REMOTE_CLIENT_NODE          $FT_REQ_REMOTE_CLIENT_NODE"
+}
 
-# source the functions in labels.sh
-. labels.sh
+# Try to determine if only host-networked pods were created.
+if [ "$FT_HOSTONLY" == unknown ]; then
+  TEST_HTTP_SERVER=`kubectl get pods | grep -o "$HTTP_SERVER_POD_NAME"`
+  if [ -z "${TEST_HTTP_SERVER}" ]; then
+    FT_HOSTONLY=true
+  else
+    FT_HOSTONLY=false
+  fi
+fi
+
+# Test for '--help'
+if [ ! -z "$1" ] ; then
+  if [ "$1" == help ] || [ "$1" == "--help" ] ; then
+    echo
+    echo "This script uses ENV Variables to control test. Here are few key ones:"
+    echo "  FT_HOSTONLY                - Only host network backed pods were launched, off by default."
+    echo "                               Used on DPUs. It is best to export this variable. test.sh and"
+    echo "                               cleanup.sh will try to detect if it was used on launch, but"
+    echo "                               false positives could occur if pods are renamed or server pod"
+    echo "                               failed to come up. Example:"
+    echo "                                 export FT_HOSTONLY=true"
+    echo "                                 ./launch.sh"
+    echo "                                 ./test.sh"
+    echo "                                 ./cleanup.sh"
+    echo "  CLEAN_ALL                  - Remove all generated files (yamls from j2, iperf logs, and"
+    echo "                               ovn-trace logs). Default is to leave in place. Example:"
+    echo "                                 CLEAN_ALL=true ./cleanup.sh"
+
+    dump-working-data
+    dump_labels
+  else
+    echo
+    echo "Unknown input, try using \"./launch.sh --help\""
+    echo
+  fi
+
+  exit 0
+fi
+
+dump-working-data
 
 # Call manage_labels() with ADD again to set flags so know what to delete
 FT_LABEL_ACTION=add
@@ -16,24 +69,26 @@ FT_NORMAL_CLIENT=false
 FT_SMARTNIC_CLIENT=false
 manage_labels
 
-# Delete normal Pods and Service
-if [ "$FT_NORMAL_CLIENT" == true ]; then
-  kubectl delete -f ./manifests/yamls/client-daemonSet.yaml
-fi
-if [ "$FT_SMARTNIC_CLIENT" == true ]; then
-  kubectl delete -f ./manifests/yamls/client-daemonSet-smartNic.yaml
-fi
+if [ "$FT_HOSTONLY" == false ]; then
+  # Delete normal Pods and Service
+  if [ "$FT_NORMAL_CLIENT" == true ]; then
+    kubectl delete -f ./manifests/yamls/client-daemonSet.yaml
+  fi
+  if [ "$FT_SMARTNIC_CLIENT" == true ]; then
+    kubectl delete -f ./manifests/yamls/client-daemonSet-smartNic.yaml
+  fi
 
-if [ "$FT_SMARTNIC_SERVER" == true ]; then
-  kubectl delete -f ./manifests/yamls/iperf-server-pod-v4-smartNic.yaml
-  kubectl delete -f ./manifests/yamls/http-server-pod-v4-smartNic.yaml
-else
-  kubectl delete -f ./manifests/yamls/iperf-server-pod-v4.yaml
-  kubectl delete -f ./manifests/yamls/http-server-pod-v4.yaml
-fi
+  if [ "$FT_SMARTNIC_SERVER" == true ]; then
+    kubectl delete -f ./manifests/yamls/iperf-server-pod-v4-smartNic.yaml
+    kubectl delete -f ./manifests/yamls/http-server-pod-v4-smartNic.yaml
+  else
+    kubectl delete -f ./manifests/yamls/iperf-server-pod-v4.yaml
+    kubectl delete -f ./manifests/yamls/http-server-pod-v4.yaml
+  fi
 
-kubectl delete -f ./manifests/yamls/svc-nodePort.yaml
-kubectl delete -f ./manifests/yamls/svc-clusterIP.yaml
+  kubectl delete -f ./manifests/yamls/svc-nodePort.yaml
+  kubectl delete -f ./manifests/yamls/svc-clusterIP.yaml
+fi
 
 
 # Delete HOST backed Pods and Service
